requirements.txt
Bottleneck==1.3.2
certifi==2021.10.8
cffi
charset-normalizer==2.0.6
cryptography
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-import-export==2.5.0
django-simpleui==2021.8.1
et-xmlfile==1.1.0
greenlet
idna==3.2
importlib-metadata
MarkupPy==1.14
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mysqlclient==1.3.14
numexpr
numpy
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.7
pandas
Pillow
pycparser
PyMySQL
python-dateutil
pytz
PyYAML==5.4.1
requests==2.26.0
six
smartchart==3.9.9.13
smartdb==0.3
SQLAlchemy
sqlparse==0.4.1
tablib==3.0.0
typing-extensions
urllib3==1.26.6
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
zipp
